pyproject.toml
[project]
name = "video-processing-backend"
version = "0.1.0"
description = "FastAPI backend for video processing"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "fastapi[standard]>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.13.0",
    "python-multipart>=0.0.6",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "aiofiles>=23.2.0",
    "python-magic>=0.4.0",
    "pillow>=10.1.0",
    "celery>=5.5.3",
    "redis>=6.4.0",
    "requests>=2.32.5",
]

